+++ TestApp.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.32929.386
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestApp", "src\TestApp\TestApp.csproj", "{9DA359F9-FF0B-45C8-83DB-61A37A2AABCF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9DA359F9-FF0B-45C8-83DB-61A37A2AABCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9DA359F9-FF0B-45C8-83DB-61A37A2AABCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9DA359F9-FF0B-45C8-83DB-61A37A2AABCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9DA359F9-FF0B-45C8-83DB-61A37A2AABCF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {06B29642-1683-4787-AD2E-E84524876B37}
+	EndGlobalSection
+EndGlobal
